turned into a template. this should work for fixed size types
